- Исправлены алгоритмы поиска релевантных рубрик, релевантных терминов. Добавлен метод генерации ПП

+++ src/main/kotlin/com/gearsy/scitechsearchengine/model/TermEmbeddingNode.kt
@@ -5,5 +5,5 @@ import com.fasterxml.jackson.annotation.JsonInclude
 @JsonInclude(JsonInclude.Include.NON_NULL)
 data class TermEmbeddingNode(
     val content: String,
-    val embedding: List<Float>
+    val embedding: List<Double>
 )
